fix(security): keep workspace config and tests merge-safe
