+++ src/main/resources/application.properties
@@ -21,10 +21,11 @@ agent.context.max-issue-comments-chars=${AGENT_CONTEXT_MAX_ISSUE_COMMENTS_CHARS:
 agent.context.max-single-issue-comment-chars=${AGENT_CONTEXT_MAX_SINGLE_ISSUE_COMMENT_CHARS:4000}
 # agent.allowed-repos=owner/repo1,owner/repo2
 
-# Agent Validation (iterative error correction)
-agent.validation.enabled=${AGENT_VALIDATION_ENABLED:true}
 # Optional host-visible root for private agent workspace parents.
 giteabot.workspaces.dir=${GITEABOT_WORKSPACES_DIR:}
+
+# Agent Validation (iterative error correction)
+agent.validation.enabled=${AGENT_VALIDATION_ENABLED:true}
 # Agent budget limits (to prevent infinite loops and excessive API usage)
 agent.budget.max-context-rounds=${AGENT_BUDGET_MAX_CONTENT_ROUNDS:10}
 agent.budget.max-rounds=${AGENT_BUDGET_MAX_ROUNDS:20}
